more gcc 4.7 fixes

+++ portod.cpp
@@ -16,9 +16,9 @@ extern "C" {
 
 using namespace std;
 
-static const int MAX_CLIENTS = 16;
-static const int MAX_CONNECTIONS = MAX_CLIENTS + 1;
-static const int POLL_TIMEOUT_MS = 1000;
+static const size_t MAX_CLIENTS = 16;
+static const size_t MAX_CONNECTIONS = MAX_CLIENTS + 1;
+static const size_t POLL_TIMEOUT_MS = 1000;
 static const string PID_FILE = "/run/portod.pid";
 
 static void RemoveRpcServer(const string &path)
